requirements.txt
numpy>=1.21.3
scipy>=1.6.1
cvxpy_base>=1.1.17
dynprog>=0.1.2
